// src/app/ihbc/models/Dependent.ts
export interface Dependent {
  userId: string;
  applicationId: string;
  dependentSeqNum: string;
  dependentLastName: string;
  dependentFirstName: string;
  dependentMiddleName: string;
  dependentGender: string;
  dependentDateOfBirth: string;
  dependentSSN: string;
  dependentTobaccoUsageInd: string;
  dependentRelationShipTypeInd: string;
  dependentRelationShipDesc: string;
  dependentChangeTypeInd: string;
  dependentMedicareInd: string;
  dependentTermDate: string;
  dependentTermReason: string;
  subscriberId: string;
  appSubmittedDate: string;
}